Added script to switch to uSv (Read Description)
I don't think this script had an even close good math or basics.
For what i understand 1080 CPM should be equal to 1 mR/hr, and 108 CPM is 1uSv/hr, so this script must be wrong and i will fix it accordingly as soon as i get a more sensitive tube.
As far as i tested this script at this state  (for my understanding) get the counts and at a 40 sec mark moves the count that should be CPM to the mR/hr so it's completely wrong?, i think it completely lacks math, will look into it.