Tube that this code is made around
Using the same tube will save you a lot of time and math.
If you are using a different tube, you will surely need to grab the specifications for it if you want to make it accurate, as soon as i can do tests and math's around this tube i will make a video explaining what you have to modify in order to get a decent reading.